+++ constants.py
@@ -1,0 +1,1 @@
+SUDOKUDOMAIN = [1,2,3,4,5,6,7,8,9] # used in sudokuGraphify and main
